backend/src/main/java/com/drumonii/loltrollbuild/security/AdminUserDetails.java
package com.drumonii.loltrollbuild.security;

import com.fasterxml.jackson.annotation.JsonIgnore;
import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.userdetails.User;
import org.springframework.security.core.userdetails.UserDetails;

import java.util.Collection;

/**
 * Admin specific user details to map only applicable properties of a {@link UserDetails}.
 */
public class AdminUserDetails extends User {

    private AdminUserDetails(String username, String password, Collection<? extends GrantedAuthority> authorities) {
        super(username, password, authorities);
    }

    private AdminUserDetails(UserDetails user) {
        this(user.getUsername(), user.getUsername(), user.getAuthorities());
    }

    public static AdminUserDetails from(UserDetails user) {
        return new AdminUserDetails(user);
    }

    @JsonIgnore
    @Override
    public String getPassword() {
        return super.getPassword();
    }

    @JsonIgnore
    @Override
    public boolean isEnabled() {
        return super.isEnabled();
    }

    @JsonIgnore
    @Override
    public boolean isAccountNonExpired() {
        return super.isAccountNonExpired();
    }

    @JsonIgnore
    @Override
    public boolean isAccountNonLocked() {
        return super.isAccountNonLocked();
    }

    @JsonIgnore
    @Override
    public boolean isCredentialsNonExpired() {
        return super.isCredentialsNonExpired();
    }

}
